Move IS_JEST_ENVIRONMENT from src/dir to src/utils
