+++ src/test/index.ts
@@ -14,4 +14,4 @@ export {
   stopThrowingReactWarnings,
 } from './react_warnings';
 export { sleep } from './sleep';
-export { IS_JEST_ENVIRONMENT } from './is_jest';
+export { IS_JEST_ENVIRONMENT } from '../utils/is_jest';
